feat: changed folder structure
