+++ frontend/src/app/(public)/(home)/layout.tsx
