+++ frontend/src/app/(public)/profile/layout.tsx
@@ -1,0 +1,27 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "@/styles/globals.css";
+import { Toaster } from "sonner";
+import Navbar from "@/components/Navbar/Navbar";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "OpenEd | Profile",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={inter.className}>
+        <Toaster richColors closeButton position="top-right" />
+        <Navbar />
+        {children}
+      </body>
+    </html>
+  );
+}
